dec, dmm, dms
