+++ lib/tools/coords/_common/formats/dec/logic/dec.dart
@@ -1,7 +1,43 @@
+import 'package:gc_wizard/tools/coords/_common/logic/coordinate_format.dart';
+import 'package:gc_wizard/tools/coords/_common/logic/coordinate_format_constants.dart';
 import 'package:gc_wizard/tools/coords/_common/logic/coordinate_parser.dart';
 import 'package:gc_wizard/tools/coords/_common/logic/coordinates.dart';
+import 'package:gc_wizard/tools/coords/_common/logic/default_coord_getter.dart';
 import 'package:gc_wizard/utils/coordinate_utils.dart';
+import 'package:gc_wizard/utils/string_utils.dart';
+import 'package:intl/intl.dart';
 import 'package:latlong2/latlong.dart';
+
+class DEC extends BaseCoordinate {
+  @override
+  CoordinateFormat get format => CoordinateFormat(CoordinateFormatKey.DEC);
+
+  DEC([double? latitude, double? longitude]) : super(latitude, longitude) {
+    this.latitude = latitude ?? defaultCoordinate.latitude;
+    this.longitude = longitude ?? defaultCoordinate.longitude;
+  }
+
+  @override
+  LatLng toLatLng() {
+    return decToLatLon(this);
+  }
+
+  static DEC fromLatLon(LatLng coord) {
+    return latLonToDEC(coord);
+  }
+
+  static DEC? parse(String input, {bool wholeString = false}) {
+    return parseDEC(input, wholeString: wholeString);
+  }
+
+  @override
+  String toString([int? precision]) {
+    precision = precision ?? 10;
+    var latFormatStr = formatStringForDecimals(decimalPrecision: precision);
+    var lonFormatStr = formatStringForDecimals(integerPrecision: 3, decimalPrecision: precision);
+    return '${NumberFormat(latFormatStr).format(latitude)}\n${NumberFormat(lonFormatStr).format(longitude)}';
+  }
+}
 
 LatLng decToLatLon(DEC dec) {
   return normalizeLatLon(dec.latitude, dec.longitude);
