+++ test/tools/coords/_common/formats/dmm/logic/dmm_test.dart
@@ -1,6 +1,6 @@
 import "package:flutter_test/flutter_test.dart";
+import 'package:gc_wizard/tools/coords/_common/formats/dmm/logic/dmm.dart';
 import 'package:gc_wizard/tools/coords/_common/logic/coordinate_format_constants.dart';
-import 'package:gc_wizard/tools/coords/_common/logic/coordinates.dart';
 import 'package:gc_wizard/utils/coordinate_utils.dart';
 import 'package:latlong2/latlong.dart';
